+++ app/Models/CdrModel.php
@@ -1,0 +1,9 @@
+<?php
+namespace App\Models;
+
+use CodeIgniter\Model;
+
+class CdrModel extends Model
+{
+    protected $table      = 'ast_cdr';
+}