+++ app/Views/layouts/sidebar.php
@@ -32,6 +32,12 @@
                     <span class="nav-text">Dashboard</span>
                 </a>                
             </li>
+            <li  class="has-sub <?= menu($menuActive,'cdr','active') ?>" >
+                <a class="sidenav-item-link" href="<?= base_url('/cdr') ?>">
+                    <i class="mdi mdi-format-list-bulleted"></i>
+                    <span class="nav-text">CDR</span>
+                </a>                
+            </li>
             <?php if (in_array('users-list', $loggedUserPermissions) && in_array('users-add', $loggedUserPermissions) ): ?>
             <li  class="has-sub <?= menu($menuActive,'users','active') ?>" >
                 <a class="sidenav-item-link" href="javascript:void(0)" data-toggle="collapse" data-target="#users" aria-expanded="false" aria-controls="users">
@@ -74,28 +80,7 @@
                         </li>
                     </div>
                 </ul>
-            </li>
-            <li  class="has-sub" >
-                <a class="sidenav-item-link" href="javascript:void(0)" data-toggle="collapse" data-target="#example" aria-expanded="false" aria-controls="example">
-                    <i class="mdi mdi-pencil-box-multiple"></i>
-                    <span class="nav-text">Example</span> <b class="caret"></b>
-                </a>                
-                <ul  class="collapse"  id="example"
-                    data-parent="#sidebar-menu">
-                    <div class="sub-menu">
-                        <li >
-                            <a class="sidenav-item-link" href="example-1.html">
-                            <span class="nav-text">Example 1</span>
-                            </a>
-                        </li>
-                        <li >
-                            <a class="sidenav-item-link" href="example-2.html">
-                            <span class="nav-text">Example 2</span>
-                            </a>
-                        </li>
-                    </div>
-                </ul>
-            </li>
+            </li>            
         </ul>
     </div>
 </div